[Discover] Create useDiscoverHistogram hook and remove old histogram dependencies from Discover
