--- src/plugins/discover/public/application/main/components/hits_counter/hits_counter.scss
@@ -1,3 +1,0 @@
-.dscHitsCounter {
-  flex-grow: 0;
-}
